+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 2.19.0 - November 28th, 2020
 - Added support for TLX (thanks [@anandoza](https://github.com/anandoza))
 - Fixed the parsing of problems and contests in the LibreOJ parsers
+- Fixed the parsing of problems in the Library Checker parser
 
 # 2.18.3 - October 18th, 2020
 - Fixed the parsing of interactive problems in the USACO parser